backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.12
pydantic-settings==2.0.3
python-multipart==0.0.6
requests==2.31.0
pyTelegramBotAPI==4.14.0
netifaces==0.11.0
python-dotenv==1.0.0